+++ freex3d/src/lib/scenegraph/Component_Time.c
@@ -37,6 +37,7 @@ X3D Time Component
 #include <libFreeWRL.h>
 
 #include "../vrml_parser/Structs.h"
+#include "../vrml_parser/CRoutes.h"
 #include "../main/headers.h"
 #include "../input/SensInterps.h"
 
@@ -134,7 +135,7 @@ void do_TimeSensorTick ( void *ptr) {
 		}
 		node->__ctflag = frac;
 
-		node->fraction_changed = frac;
+		node->fraction_changed = (float) frac;
 		MARK_EVENT (ptr, offsetof(struct X3D_TimeSensor, fraction_changed));
 
 	}
